stories: fix color/shadow stories and don't export components
