+++ src/components/styles/__stories__/Colors.stories.tsx
@@ -1,11 +1,14 @@
-import type { Meta, StoryFn } from '@storybook/react';
+import type { Meta } from '@storybook/react';
 import type { Page } from 'puppeteer';
+import styled from 'styled-components';
 
-import { ColorPalette } from 'src/components/styles/ColorPalette';
+import { VStack } from 'src/components/stack/VStack';
+import { Text } from 'src/components/text/Text';
+import { theme } from 'src/styles/constants/theme';
+import { type Color, colorContrasts, colorPrefixes } from 'src/types/Color';
 import { customSnapshotsDir } from 'src/utils/_snapshotsFolder';
 
 const StyleMeta: Meta = {
-  component: ColorPalette,
   parameters: {
     design: {
       type: 'figma',
@@ -18,10 +21,59 @@ const StyleMeta: Meta = {
       });
     },
   },
+  title: 'Color Palette',
 };
 
 export default StyleMeta;
 
-const ColorTemplate: StoryFn = () => <ColorPalette />;
+const Wrapper = styled.div`
+  display: grid;
+  grid-template-columns: repeat(auto-fit, calc(33% - ${theme.space24}));
+  gap: ${theme.space24};
+  align-items: center;
+`;
 
-export const Colors = ColorTemplate.bind({});
+const ColorWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin-bottom: ${theme.space24};
+  text-align: center;
+`;
+
+const StyledColorIntensity = styled.div<{
+  $color: Color;
+  background: Color;
+}>`
+  color: ${p => theme[p.$color]};
+  font-size: ${theme.fontSizeS};
+  padding: ${theme.space24};
+  background: ${p => theme[p.background]};
+`;
+
+export const ColorPalette = () => (
+  <Wrapper>
+    {colorPrefixes.map(color => (
+      <ColorWrapper key={color}>
+        <Text type="title">{color.toUpperCase()}</Text>
+        <VStack spacing={0}>
+          {(color !== 'gray' && color !== 'glass'
+            ? colorContrasts
+            : ['0', '50', ...colorContrasts, '1100', '1200']
+          ).map(value => {
+            const aminoColor: Color = `${color}${value}` as Color;
+            return (
+              <div key={aminoColor}>
+                <StyledColorIntensity
+                  $color={Number(value) <= 500 ? 'gray1200' : 'gray0'}
+                  background={aminoColor}
+                >
+                  <Text>{value}</Text>
+                </StyledColorIntensity>
+              </div>
+            );
+          })}
+        </VStack>
+      </ColorWrapper>
+    ))}
+  </Wrapper>
+);
